Make it easier to find this module on the Forge
This commit adds tags to the metadata.json file. The goal of that being
improved findability on the Forge.

Prior to this, if a user searched for 'atom' on the Forge, this module
would not be listed. After this commit, this module will be included in
the list of modules when searching for 'atom'.